Committer: Kevin J Baird <knowledge@storj.io>
On branch feature/visual-add-nodeid
Your branch is up-to-date with 'origin/feature/visual-add-nodeid'.

 Changes to be committed:
	modified:   app/views/overview/index.js

Adds NodeID to the visual columns in the index.

Users often ask where NodeID is, and admins provide inexact information about looking inside the config file to find it.
You can see it if you hover your mouse over the # sign, but you cannot copy it from there, making it useless.
Users can now find it listed with every node in the list, it appears to the right of the # sign.